View를 분리해 StoreUserInfo Servlet을 제거
